Got the first database driven consumer app page working
I have coded the product-details.html page to generate the product
details pulled from the database collection (products), it is not the
ideal place holder but gives an understanding of what needs to be done.

I had to change the server.js, controller.js and product-details.html
to make it work, I have also earlier updated the dbscript.txt which
when run creates the sample db which would work perfectly to show one
item in the product details page.

As a shopper when i select a category i will be able to reach this page
with the item described. I have yet to connect the category page with
the details page though.